rename_Components_to_Endpoints
